Consolidates packets
